Added Socket.io
Introduced socket.io in the system and created event update(initiate from user) and change(initiate form the server) both event will receive an object containing the index id of the tile whose colour is to be changed and the colour attribute. Local change of colour feature has been removed and now only sever can command the users to change the colour. And users can request the server to change the colour of tile in all users' canvas.